requirements.txt
gunicorn==18.0
wsgiref==0.1.2
django==1.7
mysql-python
Pillow==2.4
djangorestframework
django-filter
boto==2.19.0
django-storages==1.1.8